Fix all 33 -Wformat warnings in the db core (LP64 printf widths)
clang on FreeBSD amd64 flagged 33 -Wformat mismatches across the db core,
in two categories, both written for Win32 where long==int32:

  - 64-bit TimeT64 args printed with %lld: on LP64 TimeT64 is `long`, which
    is a distinct type from `long long` (both 64-bit, so value-correct but
    UB-per-standard and warned). Cast the arg to (long long) and keep %lld
    (matches the neighbouring 64-bit gold fields). Sites: Cache.cpp,
    ClientManager.cpp, ClientManagerPlayer.cpp.
  - 32-bit int32_t args printed with %ld: on LP64 long is 8 bytes, so the
    specifier no longer matches the 4-byte int. The values are genuinely
    32-bit, so switch the specifier to %d (matches the neighbouring %d
    columns). Sites: ClientManagerBoot.cpp, ClientManagerEventFlag.cpp,
    ClientManagerPlayer.cpp, GuildManager.cpp.

Both fixes are value-preserving and produce identical SQL text. Verified
0 -Wformat warnings on a clean FreeBSD clang build; specifiers/casts are
also correct under MSVC LLP64 (int32_t=int -> %d, TimeT64=int64_t ->
(long long)/%lld), so the Windows x64 build is unaffected. No pragma or
compiler-flag suppression used.

+++ Lead-Server-Source/db/src/Cache.cpp
@@ -95,9 +95,9 @@ void CItemCache::OnFlush()
 		{
 			iLen += snprintf(szColumns + iLen, sizeof(szColumns) - iLen, ", socket0, socket1, socket2");
 			iValueLen += snprintf(szValues + iValueLen, sizeof(szValues) - iValueLen,
-						", %lld, %lld, %lld", p->alSockets[0], p->alSockets[1], p->alSockets[2]);
+						", %lld, %lld, %lld", (long long) p->alSockets[0], (long long) p->alSockets[1], (long long) p->alSockets[2]);
 			iUpdateLen += snprintf(szUpdate + iUpdateLen, sizeof(szUpdate) - iUpdateLen,
-					", socket0=%lld, socket1=%lld, socket2=%lld", p->alSockets[0], p->alSockets[1], p->alSockets[2]);
+					", socket0=%lld, socket1=%lld, socket2=%lld", (long long) p->alSockets[0], (long long) p->alSockets[1], (long long) p->alSockets[2]);
 		}
 
 		if (isAttr)
